remove dummy data
