+++ client/components/Card/Description.tsx
@@ -3,9 +3,7 @@ function Description() {
     <div
       id="description"
       className="flex items-center text-gray-500 size-4 w-full py-4"
-    >
-      good luck runnera;lskdjfkl;asjd;flkajsdkl;ja9f890as8d0f9a8sd09f8as0d9f
-    </div>
+    ></div>
   );
 }
