+++ client/components/Card/Genre.tsx
@@ -1,7 +1,5 @@
 function Genre() {
-  return (
-    <div className="font-[vt323] text-gray-500  2 w-fit px-2 ">Shooter</div>
-  );
+  return <div className="font-[vt323] text-gray-500  2 w-fit px-2 "></div>;
 }
 
 export default Genre;
