+++ client/components/Card/Status.tsx
@@ -1,8 +1,6 @@
 function Status() {
   return (
-    <div className="font-[vt323] text-lime-500 outline outline-lime-500 outline-dashed outline-offset-2 w-fit px-2 ">
-      In-Progress
-    </div>
+    <div className="font-[vt323] text-lime-500 outline outline-lime-500 outline-dashed outline-offset-2 w-fit px-2 "></div>
   );
 }
