+++ client/components/Card/Title.tsx
@@ -1,9 +1,5 @@
 function Title() {
-  return (
-    <h1 id="title" className="font-[GlassTTY] text-lime-500">
-      Marathon
-    </h1>
-  );
+  return <h1 id="title" className="font-[GlassTTY] text-lime-500"></h1>;
 }
 
 export default Title;
